- Convert various byte counters to use size_t instead of uintmax_t

+++ include/command.h
@@ -80,7 +80,7 @@ struct Command
   unsigned int count;  /**< Number of times command used */
   unsigned int rcount;  /**< Number of times command used by server */
   unsigned int ecount;  /**< Number of times command has been issued via ENCAP */
-  uintmax_t bytes;  /**< Bytes received for this command */
+  size_t bytes;  /**< Bytes received for this command */
 
   /* handlers:
    * UNREGISTERED, CLIENT, SERVER, ENCAP, OPER, LAST
